Support Koala Feature Drop

+++ src/main/kotlin/com/pluu/plugin/toolWindow/device/DeviceToolWindowManager.kt
@@ -12,6 +12,7 @@ import com.android.tools.idea.deviceprovisioner.DeviceProvisionerService
 import com.android.tools.idea.streaming.emulator.EmulatorController
 import com.android.tools.idea.streaming.emulator.EmulatorController.ConnectionState
 import com.android.tools.idea.streaming.emulator.EmulatorController.ConnectionStateListener
+import com.android.tools.idea.streaming.emulator.RunningEmulatorCatalog
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.components.service
@@ -27,10 +28,9 @@ import com.intellij.openapi.wm.ex.ToolWindowManagerListener.ToolWindowManagerEve
 import com.intellij.openapi.wm.ex.ToolWindowManagerListener.ToolWindowManagerEventType.ShowToolWindow
 import com.intellij.ui.content.ContentFactory
 import com.intellij.util.IncorrectOperationException
-import com.pluu.plugin.toolWindow.device.controller.RunningEmulatorCatalog
 import java.awt.EventQueue
 
-private const val EMULATOR_DISCOVERY_INTERVAL_MILLIS = 1000
+private const val EMULATOR_DISCOVERY_INTERVAL_MILLIS = 1000L
 
 @Suppress("IncorrectServiceRetrieving")
 @UiThread
@@ -129,7 +129,7 @@ internal class DeviceToolWindowManager(
         }
         contentShown = false
 
-        RunningEmulatorCatalog.getInstance().addListener(this, Int.MAX_VALUE) // Don't need frequent updates.
+        RunningEmulatorCatalog.getInstance().addListener(this, Long.MAX_VALUE) // Don't need frequent updates.
         for (emulator in emulators) {
             emulator.removeConnectionStateListener(connectionStateListener)
         }
